+++ Lusas_Adapter/CRUD/Create/Line.cs
@@ -43,57 +43,9 @@ namespace BH.Adapter.Lusas
                 string supportName = "Sp" + bar.Spring.CustomData[AdapterId] + "/" + bar.Spring.Name;
                 IFAttribute lusasSupport = d_LusasData.getAttribute("Support", supportName);
                 lusasSupport.assignTo(newLine);
-                object lineXAxis = null;
-                object lineYAxis = null;
-                object lineZAxis = null;
-                object origin = null;
-
-                newLine.getAxesAtNrmCrds(0.5, ref origin, ref lineXAxis, ref lineYAxis, ref lineZAxis);
-
-                double[] localXAxis = ConvertToDouble(lineXAxis);
-                double[] localYAxis = ConvertToDouble(lineYAxis);
-                double[] localZAxis = ConvertToDouble(lineZAxis);
-
-                Point barStart = bar.StartNode.Position;
-                double[] worldXAxis = new double[] { 1, 0, 0 };
-                double[] worldYAxis = new double[] { 0, 1, 0 };
-                double[] worldZAxis = new double[] { 0, 0, 1 };
-
-                double[] barorigin = new double[] { barStart.X, barStart.Y, barStart.Z };
-
-                double[] matrixCol0 = new double[]
-                {
-                worldXAxis.Zip(localXAxis, (d1,d2) => d1 * d2).Sum(),
-                worldYAxis.Zip(localXAxis, (d1,d2) => d1 * d2).Sum(),
-                worldZAxis.Zip(localXAxis, (d1,d2) => d1 * d2).Sum(),
-                };
-
-                double[] matrixCol1 = new double[]
-                {
-                worldXAxis.Zip(localYAxis, (d1,d2) => d1 * d2).Sum(),
-                worldYAxis.Zip(localYAxis, (d1,d2) => d1 * d2).Sum(),
-                worldZAxis.Zip(localYAxis, (d1,d2) => d1 * d2).Sum(),
-                };
-
-                double[] matrixCol2 = new double[]
-                {
-                worldXAxis.Zip(localZAxis, (d1,d2) => d1 * d2).Sum(),
-                worldYAxis.Zip(localZAxis, (d1,d2) => d1 * d2).Sum(),
-                worldZAxis.Zip(localZAxis, (d1,d2) => d1 * d2).Sum(),
-                };
-
-                IFLocalCoord barLocalAxis = d_LusasData.createLocalCartesianAttr(
-                    "LocalCoordinateTest",
-                    barorigin,
-                    matrixCol0,
-                    matrixCol1,
-                    matrixCol2
-                    );
-
+                IFLocalCoord barLocalAxis = CreateLocalCoordinate(newLine);
                 barLocalAxis.assignTo(newLine);
-
             }
-
 
 
             return newLine;
