+++ Lusas_Adapter/Lusas_Adapter.csproj
@@ -91,6 +91,7 @@
     <Compile Include="CRUD\Create\Edge.cs" />
     <Compile Include="CRUD\Create\Loadcase.cs" />
     <Compile Include="CRUD\Create\LoadCombination.cs" />
+    <Compile Include="CRUD\Create\LocalCoordinate.cs" />
     <Compile Include="CRUD\Create\Material.cs" />
     <Compile Include="CRUD\Create\Support.cs" />
     <Compile Include="CRUD\Create\Surface.cs" />
